requirements.txt
accelerate==0.23.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
certifi==2025.10.5
charset-normalizer==3.4.3
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
diffusers==0.30.1
executing==2.2.1
filelock==3.13.1
fonttools==4.60.1
fsspec==2024.6.1
hf-xet==1.1.10
huggingface-hub==0.35.3
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.7.0
iopath==0.1.10
ipykernel==6.30.1
ipympl==0.9.8
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
opencv-python==4.12.0.88
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.4.0
portalocker==3.2.0
prompt_toolkit==3.0.52
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.9.18
requests==2.32.5
safetensors==0.3.1
-e git+https://github.com/facebookresearch/sam2.git@5660069ba6d6e0940a2d0700d1eb90d842792779#egg=SAM_2
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tokenizers==0.13.3
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.33.2
triton==3.3.1
typing_extensions==4.12.2
urllib3==2.5.0
wcwidth==0.2.14
widgetsnbextension==4.0.14
xformers==0.0.31.post1
zipp==3.23.0
